requirements.txt
absl-py==2.1.0
accelerate==0.34.2
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
bleach==6.1.0
blinker==1.8.2
blip-ci==0.0.5
blis==0.7.11
braceexpand==0.1.7
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==2.1.1
click==8.1.7
cloudpathlib==0.19.0
confection==0.1.5
contexttimer==0.3.3
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.8
datasets==3.0.0
decorator==5.1.1
decord==0.6.0
diffusers==0.30.3
dill==0.3.8
distlib==0.3.8
einops==0.8.0
exceptiongroup==1.2.2
executing==2.1.0
fairscale==0.4.4
filelock==3.13.1
fire==0.6.0
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.2.3
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.25.0
identify==2.6.1
idna==3.4
imageio==2.35.1
importlib_metadata==8.5.0
iopath==0.1.10
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kaggle==1.6.17
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
lazy_loader==0.4
lightning-utilities==0.11.7
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.10
narwhals==1.8.4
networkx==3.2.1
nodeenv==1.9.1
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
omegaconf==2.3.0
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.3.6
plotly==5.24.1
portalocker==2.10.1
pre-commit==3.8.0
preshed==3.0.9
prompt_toolkit==3.0.48
protobuf==5.28.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pycocoevalcap==1.2
pycocotools==2.0.8
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-magic==0.4.27
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
safetensors==0.4.5
salesforce-lavis==1.0.2
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
streamlit==1.38.0
sympy==1.12
tenacity==8.5.0
tensorboardX==2.6.2.2
termcolor==2.4.0
text-unidecode==1.3
thinc==8.2.5
threadpoolctl==3.5.0
tifffile==2024.9.20
timm==0.4.12
tokenizers==0.13.3
toml==0.10.2
torch==2.4.0+cu118
torchaudio==2.4.0+cu118
torchmetrics==1.4.2
torchvision==0.19.0+cu118
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.26.1
triton==3.0.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.13
virtualenv==20.26.6
wasabi==1.1.3
watchdog==4.0.2
wcwidth==0.2.13
weasel==0.4.1
webdataset==0.2.100
webencodings==0.5.1
wrapt==1.16.0
xxhash==3.5.0
yarl==1.12.1
zipp==3.20.2
